projection/layout: Rename _helper to _base
This is what they are, the rename makes finding the base a lot easier.
