+++ OZprivate/rawJS/OZTreeModule/src/projection/layout/AT/branch_layout.js
@@ -1,4 +1,4 @@
-import BranchLayoutBase from '../branch_layout_helper';
+import BranchLayoutBase from '../branch_layout_base';
 
 class ATBranchLayout extends BranchLayoutBase {
 }
