+++ OZprivate/rawJS/OZTreeModule/src/projection/layout/life/branch_layout.js
@@ -1,3 +1,3 @@
-import BranchLayoutBase from '../branch_layout_helper';
+import BranchLayoutBase from '../branch_layout_base';
 
 export default BranchLayoutBase;
