+++ OZprivate/rawJS/OZTreeModule/src/projection/layout/life/leaf_layout.js
@@ -1,3 +1,3 @@
-import LeafLayoutBase from '../leaf_layout_helper';
+import LeafLayoutBase from '../leaf_layout_base';
 
 export default LeafLayoutBase;
