+++ OZprivate/rawJS/OZTreeModule/src/projection/layout/life/node_layout.js
@@ -1,3 +1,3 @@
-import NodeLayoutBase from '../node_layout_helper';
+import NodeLayoutBase from '../node_layout_base';
 
 export default NodeLayoutBase;
