+++ OZprivate/rawJS/OZTreeModule/src/projection/layout/otop/branch_layout.js
@@ -1,4 +1,4 @@
-import BranchLayoutBase from '../branch_layout_helper';
+import BranchLayoutBase from '../branch_layout_base';
 import BezierShape from '../../shapes/bezier_shape';
 import ArcShape from '../../shapes/arc_shape';
 import {color_theme} from '../../../themes/color_theme';
