+++ OZprivate/rawJS/OZTreeModule/src/projection/layout/polytomy2/leaf_layout.js
@@ -1,4 +1,4 @@
-import LeafLayoutBase from '../leaf_layout_helper';
+import LeafLayoutBase from '../leaf_layout_base';
