+++ OZprivate/rawJS/OZTreeModule/src/projection/layout/polytomy2/node_layout.js
@@ -1,4 +1,4 @@
-import NodeLayoutBase from '../node_layout_helper';
+import NodeLayoutBase from '../node_layout_base';
 import config from '../../../global_config';
 
 class PolytomyNodeLayout extends NodeLayoutBase {
