+++ OZprivate/rawJS/OZTreeModule/src/projection/layout/polytomy2/signpost_layout.js
@@ -1,5 +1,5 @@
 // this overwrites the default behaviour of signpost_layout
 
-import SignpostLayoutBase from '../signpost_layout_helper';
+import SignpostLayoutBase from '../signpost_layout_base';
 
 export default SignpostLayoutBase;
